requirements.txt
fastapi
uvicorn[standard]
python-dotenv
python-jose[cryptography]
sqlalchemy
langchain<1.0
langchain-core<1.0
langchain-openai<1.0
langchain-groq<1.0
langchain-community<1.0
langgraph<1.0
langchain-tavily<1.0
pydantic
paddleocr
paddlepaddle
opencv-python<4.12
numpy<2.0
